ReachabilitySwift conflicts with Reachability (close #437)
PR: #475

* Fix issue when Podfile without "use_frameworks!"

Using Podfile without "use_frameworks!" can generate issues.
In that case `#import <SnowplowTracker/SnowplowTracker-Swift.h>` raises the error "file not found".
Indeed the file doesn't exist and we have to use `#import "SnowplowTracker-Swift.h"`.
This is a temporary workaround in order to build the tracker with and without "use_frameworks!" directive on Podfile.

* Remove ReachabilitySwift dependency

We don't need ReachabilitySwift dependency because we don't want swift classes inside the objc tracker. It can cause issues on compilation when the tracker is in edge cases. Also the introduction of swift classes causes the requirement of bridging headers, particularly annoying when used on React Native.

* Check network status without use of ReachabilitySwift dependency

* Don't prevent events sending

We can never trust of Reachability code before to send events. Apple's Reachability is intended as diagnostic if a connection doesn't work but it can't be used to prevent a connection because it can report false negatives.

* Remove reference to isOnline as it has been removed

* Fix building issues on macOS target

* Update podspec with DEFINES_MODULE set to YES

This setting is needed by Cocoapods as they declared in the doc:
"you are able to opt into stricter header search paths (and module map generation for Objective-C pods). As a pod author, you can add 'DEFINES_MODULE' => 'YES' to your pod_target_xcconfig. Alternatively, in your Podfile you can add use_modular_headers! to enable the stricter search paths and module map generation for all of your pods, or you can add :modular_headers => true to a single pod declaration to enable for only that pod."

* Remove dependency of ReachabilitySwift from Cartfile

* Update demo apps

* Update travis script

- Added bash scripts in order to make it testable locally (without travis)
- Added multiple Podfiles in order to test different settings