Add Watchkit support (close #465)
PR #478

* Adds WatchKit support via CocoaPods (#465)

* Fixes missing file for watchOS

* Change watchOS product name

* Adds modual map for watchos

* Adds carthage watch example

* Adds cocoapods watch example

* Update CocoaPods demo to include older watchOS version

I noticed that apps using this library could not launch on older watchOS version. This adds a demo project to reproduce this,

* Fix crash on < watchOS 6

This fixes the crash mentioned in the prev. commit. It removes the frameworks for watchOS from the Podspec.

* Changed button name

* Update podspec for watchOS target

Updated podspec excluding the new Reachability feature added when we got rid of the third part dependency Reachability.swift

* Update travis scripts

* Pod update

Co-authored-by: leoAsana <56038088+leoAsana@users.noreply.github.com>