Import corrections from Android tracker (close #563) (#565)
* Minor GDPR configuration and controller changes

Adapted to Android implementation

* Minor Global Contexts configuration and controller changes

Adapted from Android implementation

* Fix ECommerce event

Adapted from Android implementation

* Rename emitThreadPoolSize to threadPoolSize

* Rename SPUnstructured to SPSelfDescribing

* Fix EcommerceItem and ConsentDocument initialisation

* Remove trackSelfDescribing method

Because track can do the same with a SelfDescribing event

* Rename Controller to ControllerImpl and Controlling to Controller

Adapted from Android implementation

* Add script to copy public headers

We need to keep all the public headers together in the root directory but we don’t want to lose the ability to structure our code in directories. The solution is to keep the sources in different directories and copying in the root directory the public headers only.

* Fix header not found on demo app importing with SPM

* Fix flaky tests

* Fix podspec

It imported public headers twice

* Remove Protocol argument from tracker setup

The endpoint field was a url without schema (protocol). Now it needs the protocol specified, otherwise it is considered https.

* Rename SPRequestOptions to SPHttpMethod

* Make platformContext and geoLocationContext settings editable at runtime

* Fix test after renaming of SPRequestOptions to SPHttpMethod

* Set SPSnowplow as entry point to setup the tracker

Previously SPTracker was the entry point to setup the tracker, now deprecated.

* Fix method signature on ServiceProvider header

* Fix method signature on ServiceProvider header