Fix contentAvailable value type
ObjC doesn't have boolean objects (only primitive types). We use NSNumber to represent booleans. This can cause troubles with the JSON serialisation as it doesn't know how to represent the value (number or boolean). Getting the `boolValue` of a NSNumber we can force the JSON serialisation to treat the value as a boolean.