Add session callback (close #664)
PR: #668

* Add configuration for session callback

* Rename onSessionUpdate to onSessionStateUpdate

* Substitute session properties with session state

* Fix test error with [SPUtilities getLanguage]

Probably the error has been raised by the new Xcode version 13.2.1.
We just care to have the language string. We don't need the specific language.
It can be a problem when the test runs on a device/emulator set with a different language
(example: it-US)

* Fix error onSessionStateUpdate not correctly updated

* Publish SPSessionState.h as it's used in the session callback

* Fix typo in SessionConfiguration builder

* Remove null values from dictionary to persist

* Execute callback in separate thread

* Fix session persistence broken because of plist serialization

SPDataPersistence uses ObjC dictionary serialization which requires plist data type. Unfortunately NSNull is not a plist data type and can't be serialized.
To store the session we need to remove key/values that have a NSNull value.

* Remove Session.getSessionIndex to use same approach of getSessionId

* Remove local sessionIndex and improve userID management

* Fix import of stored session with legacy format

* Fix headers on new source-files

* Fix SessionConfiguration doesn't pass onSessionUpdate callback in its clone

* Update explanation of SessionUserID and InstallationUserID